TEI2TXTConverter and TEI2TXTConverterTest
Currently failing test to convert a minimal tei to txt using XSL stylesheet from
https://github.com/TEIC/Stylesheets/blob/dev/txt/tei-to-text.xsl

This XSL had the import statement commented out since this was needed for it to correctly run
within the norma framework. This may be the reason for the failing test.

Conversion implemented and close to successful using TransformerWrapper.

Also give some human readable and close to correct results when manually run on
real TEIs from GROBID